put endpoint version info into existing extension
