--- gradle-conjure-api/src/main/java/com/palantir/gradle/conjure/api/ConjureProductDependenciesExtension.java
@@ -1,40 +1,0 @@
-/*
- * (c) Copyright 2018 Palantir Technologies Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.palantir.gradle.conjure.api;
-
-import groovy.lang.Closure;
-import groovy.lang.DelegatesTo;
-import java.util.HashSet;
-import java.util.Set;
-
-public class ConjureProductDependenciesExtension {
-
-    public static final String EXTENSION_NAME = "serviceDependencies";
-
-    private final Set<ServiceDependency> productDependencies = new HashSet<>();
-
-    public final Set<ServiceDependency> getProductDependencies() {
-        return productDependencies;
-    }
-
-    public final void serviceDependency(@DelegatesTo(ServiceDependency.class) Closure<ServiceDependency> closure) {
-        ServiceDependency serviceDependency = new ServiceDependency();
-        closure.setDelegate(serviceDependency);
-        closure.call();
-        productDependencies.add(serviceDependency);
-    }
-}
