+++ gradle-conjure/src/main/java/com/palantir/gradle/conjure/ConjurePlugin.java
@@ -23,7 +23,6 @@ import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
 import com.palantir.gradle.conjure.api.ConjureExtension;
-import com.palantir.gradle.conjure.api.ConjureProductDependenciesExtension;
 import com.palantir.gradle.conjure.api.GeneratorOptions;
 import java.io.File;
 import java.lang.reflect.InvocationTargetException;
@@ -51,11 +50,9 @@ import org.gradle.api.logging.Logger;
 import org.gradle.api.logging.Logging;
 import org.gradle.api.plugins.BasePlugin;
 import org.gradle.api.plugins.JavaLibraryPlugin;
-import org.gradle.api.plugins.JavaPlugin;
 import org.gradle.api.plugins.JavaPluginConvention;
 import org.gradle.api.tasks.Exec;
 import org.gradle.api.tasks.TaskProvider;
-import org.gradle.jvm.tasks.Jar;
 import org.gradle.plugins.ide.eclipse.EclipsePlugin;
 import org.gradle.plugins.ide.idea.IdeaPlugin;
 import org.gradle.plugins.ide.idea.model.IdeaModule;
@@ -156,12 +153,6 @@ public final class ConjurePlugin implements Plugin<Project> {
 
         ExtractExecutableTask extractJavaTask = ExtractConjurePlugin.applyConjureJava(project);
 
-        EndpointMinimumVersionsExtension minimumVersionsExtension = project.getExtensions()
-                .create(
-                        EndpointMinimumVersionsExtension.EXTENSION_NAME,
-                        EndpointMinimumVersionsExtension.class,
-                        project);
-
         Map<String, Consumer<Project>> configs = ImmutableMap.<String, Consumer<Project>>builder()
                 .put(JAVA_OBJECTS_SUFFIX, (Consumer<Project>) ConjurePlugin::setupObjectsProject)
                 .put(JAVA_DIALOGUE_SUFFIX, (Consumer<Project>) ConjurePlugin::setupDialogueProject)
@@ -187,7 +178,6 @@ public final class ConjurePlugin implements Plugin<Project> {
                 compileIrTask,
                 productDependencyExt,
                 extractJavaTask,
-                minimumVersionsExtension,
                 config));
     }
 
@@ -199,7 +189,6 @@ public final class ConjurePlugin implements Plugin<Project> {
             TaskProvider<?> compileIrTask,
             ConjureProductDependenciesExtension productDependencyExt,
             ExtractExecutableTask extractJavaTask,
-            EndpointMinimumVersionsExtension minimumVersionsExtension,
             Consumer<Project> extraConfig) {
         String projectName = getDerivedProjectName(parentProject, projectSuffix);
         if (!derivedProjectExists(parentProject, projectName)) {
@@ -239,30 +228,11 @@ public final class ConjurePlugin implements Plugin<Project> {
             cleanTask.dependsOn(parentProject.getTasks().findByName("cleanCompileConjure" + upperSuffix));
             if (isNotObjectsProject) {
                 ConjureJavaServiceDependencies.configureJavaServiceDependencies(subproj, productDependencyExt);
-                configureEndpointMinimumVersionTask(subproj, minimumVersionsExtension);
                 subproj.getDependencies().add("api", findDerivedProject(parentProject, objectsProjectName));
             }
             if (extraConfig != null) {
                 extraConfig.accept(subproj);
             }
-        });
-    }
-
-    private static void configureEndpointMinimumVersionTask(Project project, EndpointMinimumVersionsExtension ext) {
-        project.getPluginManager().withPlugin("java", _plugin -> {
-            TaskProvider<ConfigureEndpointMinimumVersionsTask> configureEndpointVersionsTask = project.getTasks()
-                    .register(
-                            "configureEndpointMinimumVersions",
-                            ConfigureEndpointMinimumVersionsTask.class,
-                            cmt -> cmt.getVersions().set(ext.getEndpointVersions()));
-
-            // Ensure that the jar task depends on this wiring task
-            project.getTasks()
-                    .withType(Jar.class)
-                    .named(JavaPlugin.JAR_TASK_NAME)
-                    .configure(jar -> {
-                        jar.dependsOn(configureEndpointVersionsTask);
-                    });
         });
     }
